refactor: consolidate facade re-export modules (v2, rebased on #526)
Remove 4 subsystem facades and strip re-exports from tab-facade.js.
Consumers now import directly from source modules.

Closes #462
